requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
ultralytics>=8.3.50
opencv-python-headless>=4.10.0
Pillow>=11.0.0
chromadb>=0.6.0
sentence-transformers>=3.3.0
google-generativeai>=0.8.0
supabase>=2.11.0
python-dotenv>=1.0.1
numpy>=2.1.0
